[vimrc] remove overlap between Rconfig and Rcontroller
In theory, a mapping to directly to a rails configfile sounds great.
In practice, :e co<TAB> is faster and better, because you get
incremental (per directory) tab-completion.

The main reason to remove this is - as stated in the headline - the
overlapping same beginning of the command with :Rcontroller. After all,
I only type :Rco<TAB>, which is not enough...